Finished BruteCollinearPoints Class with testing in sample client

+++ src/Point.java
@@ -50,7 +50,7 @@ public class Point implements Comparable<Point> {
         if (this.x == that.x && this.y == that.y) return Double.NEGATIVE_INFINITY;
         else if (this.y == that.y) return 0.0;
         else if (this.x == that.x) return Double.POSITIVE_INFINITY;
-        else return (that.y - this.y) / (that.x - this.x);
+        else return (double) (that.y - this.y) / (that.x - this.x);
     }
 
     /**
